Add gnomeos sdk and platform images
This will be used as the base for the gnome-os application work.
Its a lot more minimal and controlled than the other bases, as
we need to be very careful about what ends up in the platform.

In particular, there need not be any system management tools in there,
as this will not be used to boot systems.