Clean up platform/sdk tasks
This removes alsa for now as we need to look into audio in more detail.
It also adds some devel headers for libraries that are already in sdk
and ensures they are also in platform (popt, lzma, archive, gmp).

We also add openssl-dev to sdk and libssl to platform. We now have all
ssl libs, but I think its better to have them in the platform than
bundled, because these are risky libs. Also, we already have part (libcrypto)
in platform.